requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
ultralytics
opencv-python-headless
kafka-python
psycopg2-binary
streamlit
pandas
lap>=0.5.12